Refactor enumerator tests: - Move to the 'Enumerators' directory - Split into separate files - Remove redundant type parameters
