+++ src/EventStore.Core.Tests/Services/Transport/Enumerators/Enumerator.AllSubscription.Tests.cs
@@ -1,0 +1,104 @@
+using System;
+using System.Collections.Generic;
+using System.Threading;
+using System.Threading.Tasks;
+using EventStore.Core.Bus;
+using EventStore.Core.Data;
+using EventStore.Core.Services.Storage.ReaderIndex;
+using EventStore.Core.Services.Transport.Common;
+using EventStore.Core.Services.Transport.Enumerators;
+using EventStore.Core.Services.UserManagement;
+using EventStore.Core.Tests.Helpers;
+using NUnit.Framework;
+
+namespace EventStore.Core.Tests.Services.Transport.Enumerators;
+
+[TestFixture]
+public partial class EnumeratorTests {
+	private static EnumeratorWrapper CreateAllSubscription(
+		IPublisher publisher,
+		Position startPosition) {
+
+		return new EnumeratorWrapper(new Enumerator.AllSubscription(
+			bus: publisher,
+			expiryStrategy: new DefaultExpiryStrategy(),
+			startPosition: startPosition,
+			resolveLinks: false,
+			user: SystemAccounts.System,
+			requiresLeader: false,
+			cancellationToken: CancellationToken.None));
+	}
+
+	[TestFixture(typeof(LogFormat.V2), typeof(string))]
+	[TestFixture(typeof(LogFormat.V3), typeof(uint))]
+	public class subscribe_all_from_start<TLogFormat, TStreamId> : TestFixtureWithExistingEvents<TLogFormat, TStreamId> {
+		private readonly List<Guid> _eventIds = new();
+
+		protected override void Given() {
+			EnableReadAll();
+			_eventIds.Add(WriteEvent("test-stream", "type1", "{}", "{Data: 1}").Item1.EventId);
+			_eventIds.Add(WriteEvent("test-stream", "type2", "{}", "{Data: 2}").Item1.EventId);
+			_eventIds.Add(WriteEvent("test-stream", "type3", "{}", "{Data: 3}").Item1.EventId);
+		}
+
+		[Test]
+		public async Task should_receive_live_caught_up_message_after_reading_existing_events() {
+			await using var sub = CreateAllSubscription(_publisher, Position.Start);
+
+			Assert.True(await sub.GetNext() is SubscriptionConfirmation);
+			Assert.AreEqual(_eventIds[0], ((Event)await sub.GetNext()).Id);
+			Assert.AreEqual(_eventIds[1], ((Event)await sub.GetNext()).Id);
+			Assert.AreEqual(_eventIds[2], ((Event)await sub.GetNext()).Id);
+			Assert.True(await sub.GetNext() is CaughtUp);
+		}
+	}
+
+	[TestFixture(typeof(LogFormat.V2), typeof(string))]
+	[TestFixture(typeof(LogFormat.V3), typeof(uint))]
+	public class subscribe_all_from_end<TLogFormat, TStreamId> : TestFixtureWithExistingEvents<TLogFormat, TStreamId> {
+		protected override void Given() {
+			EnableReadAll();
+			WriteEvent("test-stream", "type1", "{}", "{Data: 1}");
+			WriteEvent("test-stream", "type2", "{}", "{Data: 2}");
+			WriteEvent("test-stream", "type3", "{}", "{Data: 3}");
+		}
+
+		[Test]
+		public async Task should_receive_live_caught_up_message_immediately() {
+			await using var sub = CreateAllSubscription(_publisher, Position.End);
+
+			Assert.True(await sub.GetNext() is SubscriptionConfirmation);
+			Assert.True(await sub.GetNext() is CaughtUp);
+		}
+	}
+
+	[TestFixture(typeof(LogFormat.V2), typeof(string))]
+	[TestFixture(typeof(LogFormat.V3), typeof(uint))]
+	public class subscribe_all_from_position<TLogFormat, TStreamId> : TestFixtureWithExistingEvents<TLogFormat, TStreamId> {
+		private readonly List<Guid> _eventIds = new();
+		private TFPos _subscribeFrom;
+
+		protected override void Given() {
+			EnableReadAll();
+			WriteEvent("test-stream", "type1", "{}", "{Data: 1}");
+			WriteEvent("test-stream", "type2", "{}", "{Data: 2}");
+			(_, _subscribeFrom) = WriteEvent("test-stream", "type3", "{}", "{Data: 3}");
+			_eventIds.Add(WriteEvent("test-stream", "type4", "{}", "{Data: 4}").Item1.EventId);
+			_eventIds.Add(WriteEvent("test-stream", "type5", "{}", "{Data: 5}").Item1.EventId);
+			_eventIds.Add(WriteEvent("test-stream", "type6", "{}", "{Data: 6}").Item1.EventId);
+		}
+
+		[Test]
+		public async Task should_receive_events_after_start_position() {
+			await using var sub = CreateAllSubscription(
+				_publisher,
+				new Position((ulong)_subscribeFrom.CommitPosition, (ulong)_subscribeFrom.PreparePosition));
+
+			Assert.True(await sub.GetNext() is SubscriptionConfirmation);
+			Assert.AreEqual(_eventIds[0], ((Event)await sub.GetNext()).Id);
+			Assert.AreEqual(_eventIds[1], ((Event)await sub.GetNext()).Id);
+			Assert.AreEqual(_eventIds[2], ((Event)await sub.GetNext()).Id);
+			Assert.True(await sub.GetNext() is CaughtUp);
+		}
+	}
+}
